admin/kafka: split out admin kafka routes
