+++ src/v/redpanda/CMakeLists.txt
@@ -49,6 +49,7 @@ v_cc_library(
     admin/recovery.cc
     admin/transaction.cc
     admin/usage.cc
+    admin/kafka.cc
     cli_parser.cc
     application.cc
     monitor_unsafe_log_flag.cc
